@@ -1,39 +1,40 @@
 {
   "name": "vue-template-test",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
     "@fortawesome/fontawesome-free-brands": "^5.1.0-3",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/pro-light-svg-icons": "^5.15.4",
-    "@fortawesome/pro-solid-svg-icons": "^5.15.4",
-    "@fortawesome/vue-fontawesome": "^2.0.6",
+    "@fortawesome/fontawesome-svg-core": "^6.1.2",
+    "@fortawesome/pro-light-svg-icons": "^6.1.2",
+    "@fortawesome/pro-solid-svg-icons": "^6.1.2",
+    "@fortawesome/vue-fontawesome": "^3",
     "@lu.se/vue-template": "^1.14.1",
-    "bootstrap-vue": "^2.21.0",
-    "core-js": "^3.20.3",
-    "vue": "^2.6.14",
-    "vue-i18n": "^8.26.8",
-    "vue-router": "^3.5.3"
+    "@popperjs/core": "^2.11.5",
+    "bootstrap": "^5.2.0",
+    "vue": "^3.2.37",
+    "vue-i18n": "^9.2.0",
+    "vue-router": "^4.1.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.5.15",
-    "@vue/cli-plugin-eslint": "^4.5.15",
-    "@vue/cli-service": "^4.5.15",
+    "@babel/core": "^7.18.10",
+    "@babel/eslint-parser": "^7.18.9",
+    "@vitejs/plugin-vue": "^3.0.1",
     "babel-eslint": "^10.0.1",
-    "eslint": "^7.0.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.25.4",
+    "eslint": "^8.21.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-n": "^15.2.4",
     "eslint-plugin-no-autofix": "^1.2.3",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.0.0",
+    "eslint-plugin-promise": "^6.0.0",
     "eslint-plugin-standard": "^5.0.0",
-    "eslint-plugin-vue": "^7.0.0",
-    "standard": "^16.0.4",
-    "vue-template-compiler": "^2.6.14"
+    "eslint-plugin-vue": "^9.3.0",
+    "standard": "^17.0.0",
+    "vite": "^3.0.4"
   }
 }
