@@ -3,36 +3,36 @@ export default {
     admin: 'Admin',
     google: 'Google',
     home: 'Hem',
-    home_title: 'Test | Hem',
+    home_title: 'Test - Hem',
     lu: 'LU',
     mail_only_for_support: 'Endast support, ej försäljning eller personlig kontakt',
     page1: 'Sidan 1',
     page2: 'Sidan 2',
     page3: 'Sidan 3',
     page4: 'Sidan 4',
-    page1_title: 'Testa | Sidan 1',
-    page2_title: 'Testa | Sidan 2',
-    page3_title: 'Testa | Sidan 3',
-    page4_title: 'Testa | Sidan 4',
+    page1_title: 'Testa - Sidan 1',
+    page2_title: 'Testa - Sidan 2',
+    page3_title: 'Testa - Sidan 3',
+    page4_title: 'Testa - Sidan 4',
     start: 'Start',
-    start_title: 'Testa | Start'
+    start_title: 'Testa - Start'
   },
   en: {
     admin: 'Admin',
     google: 'Google',
     home: 'Home',
-    home_title: 'Testa | Home',
+    home_title: 'Testa - Home',
     lu: 'LU',
     mail_only_for_support: 'Only for support, not sales or personal contact',
     page1: 'Page 1',
     page2: 'Page 2',
     page3: 'Page 3',
     page4: 'Page 4',
-    page1_title: 'Test | Page 1',
-    page2_title: 'Test | Page 2',
-    page3_title: 'Test | Page 3',
-    page4_title: 'Test | Page 4',
+    page1_title: 'Test - Page 1',
+    page2_title: 'Test - Page 2',
+    page3_title: 'Test - Page 3',
+    page4_title: 'Test - Page 4',
     start: 'Start',
-    start_title: 'Test | Start'
+    start_title: 'Test - Start'
   }
 }
